python_client/pyproject.toml
[project]
name = "python-client"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12,<3.13"
dependencies = [
    "grpcio>=1.67.1",
    "grpcio-tools>=1.67.1",
    "ruff>=0.7.2",
]

[dependency-groups]
dev = [
    "ruff>=0.7.2",
]



[tool.ruff]
exclude = [
    ".venv", "protogen"
]
line-length = 120
target-version = "py312"


[tool.ruff.format]
# Like Black, use double quotes for strings.
quote-style = "double"

# Like Black, indent with spaces, rather than tabs.
indent-style = "space"

# Like Black, respect magic trailing commas.
skip-magic-trailing-comma = false

# Like Black, automatically detect the appropriate line ending.
line-ending = "auto"
docstring-code-format = true
